Update Layout.php
Nova's DateTime field (at least the one in Nova v5) uses a static anonymous function to define how to resolve the data to a string. This causes problems in the cloneField (and the dependent duplicateAndHydrate) method in your code, which binds a few of the field callbacks back to the newly cloned field instance. The specific error is:

Cannot bind an instance to a static closure
I initially implemented a fix using ReflectionFunction to detect if a function is static, but settled for a try-catch just in case reflection causes some sort of performance issue at low numbers.